+++ bootstrap/plugin.php
@@ -14,7 +14,6 @@ use BarefootEngine\Properties\Property_Metaboxes;
 use BarefootEngine\Properties\Property_Post_Type;
 use BarefootEngine\Properties\Property_Taxonomies;
 use BarefootEngine\REST\Api_Integration_Controller;
-use BarefootEngine\REST\General_Settings_Controller;
 use BarefootEngine\REST\Property_Availability_Controller;
 use BarefootEngine\REST\Property_Booking_Checkout_Controller;
 use BarefootEngine\REST\Property_Booking_Controller;
@@ -22,7 +21,6 @@ use BarefootEngine\REST\Properties_Controller;
 use BarefootEngine\REST\Updates_Controller;
 use BarefootEngine\Services\Api_Integration_Settings;
 use BarefootEngine\Services\Barefoot_Api_Client;
-use BarefootEngine\Services\General_Settings;
 use BarefootEngine\Services\Property_Sync_Service;
 use BarefootEngine\Services\Updates_Service;
 use BarefootEngine\Widgets\Booking\Booking_Widget_Preset_Registry;
@@ -119,8 +117,6 @@ class Plugin
         $settings = $this->get_api_settings();
         $api_client = $this->get_api_client();
         $controller = new Api_Integration_Controller($settings, $api_client);
-        $general_settings = new General_Settings();
-        $general_controller = new General_Settings_Controller($general_settings);
         $updates_service = new Updates_Service();
         $updates_controller = new Updates_Controller($updates_service);
         $delta_refresh_service = $this->get_property_delta_refresh_service();
@@ -130,7 +126,6 @@ class Plugin
         $booking_checkout_controller = new Property_Booking_Checkout_Controller($this->get_property_booking_checkout_service());
 
         $this->loader->add_action('rest_api_init', $controller, 'register_routes', 10, 0);
-        $this->loader->add_action('rest_api_init', $general_controller, 'register_routes', 10, 0);
         $this->loader->add_action('rest_api_init', $updates_controller, 'register_routes', 10, 0);
         $this->loader->add_action('rest_api_init', $properties_controller, 'register_routes', 10, 0);
         $this->loader->add_action('rest_api_init', $availability_controller, 'register_routes', 10, 0);